proyectoencuestas/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
choice==0.1
crispy-bootstrap5==0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-admin-interface==0.19.0
django-colorfield==0.6.3
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==2.8.0
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
filelock==3.4.2
gunicorn==20.1.0
MarkupPy==1.14
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.1
Pillow==9.1.0
pipenv==2022.1.8
platformdirs==2.5.0
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
